Added commands to aggregate data.

+++ src/TheFipster.Aviation.CoreCli/JsonWriter.cs
@@ -4,25 +4,11 @@ namespace TheFipster.Aviation.CoreCli
 {
     public class JsonWriter<T>
     {
-
-
         public void Write(T data, string fileType, string from, string? to = null)
         {
             var location = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
 
-            var file = string.IsNullOrEmpty(to)
-                ? $"{@from} - {fileType}.json"
-                : $"{@from} - {to} - {fileType}.json";
-
-            var path = Path.Combine(location, "Aviation", "output");
-
-            if (!Directory.Exists(path))
-                Directory.CreateDirectory(path);
-
-            path = Path.Combine(path, file);
-
-            var json = JsonSerializer.Serialize(data);
-            File.WriteAllText(path, json);
+            Write(location, data, fileType, from, to);
         }
 
         public void Write(string flightPath, T data, string fileType, string from, string? to = null)
@@ -31,7 +17,15 @@ namespace TheFipster.Aviation.CoreCli
                 ? $"{@from} - {fileType}.json"
                 : $"{@from} - {to} - {fileType}.json";
 
-            var path = Path.Combine(flightPath, file);
+            Write(flightPath, data, file);
+        }
+
+        public void Write(string flightPath, T data, string filename, bool overwrite = false)
+        {
+            var path = Path.Combine(flightPath, filename);
+
+            if (File.Exists(path) && !overwrite)
+                return;
 
             var json = JsonSerializer.Serialize(data);
             File.WriteAllText(path, json);
